Moved visualisation and csv output into seperate files
